requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
frozenlist==1.4.1
future==1.0.0
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==8.2.0
importlib-resources==6.4.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.2
jupyter-core==5.7.2
jupyterlab-pygments==0.3.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.9
langchain-mongodb==0.1.6
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langsmith==0.1.81
line-bot-sdk==3.11.0
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.24.4
openai==1.35.1
orjson==3.10.5
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.2
prompt-toolkit==3.0.47
proto-plus==1.23.0
protobuf==5.27.1
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.7.4
pydantic-core==2.18.4
pygments==2.18.0
pymongo==4.7.3
PyMuPDF==1.24.9
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rpds-py==0.18.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
tenacity==8.4.1
tiktoken==0.7.0
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
werkzeug==3.0.3
wrapt==1.16.0
XlsxWriter==3.2.0
yarg==0.1.9
yarl==1.9.4
zipp==3.19.2
pdfplumber==0.11.2
openpyxl==3.1.5
bcrypt==4.2.0
mysql-connector-python==9.0.0